SITE-6395 -- update example application to support creating new foods with full nutritional information

+++ fitbit4j/src/main/java/com/fitbit/api/client/FitbitApiClientAgent.java
@@ -325,7 +325,7 @@ public class FitbitApiClientAgent extends FitbitAPIClientSupport implements Seri
         return createFood(localUser, name, description, defaultFoodMeasurementUnitId, defaultServingSize, formType, nutritionalValuesEntry);
     }
 
-    public Food createFood(LocalUserDetail localUser, String name, String description, float defaultFoodMeasurementUnitId,
+    public Food createFood(LocalUserDetail localUser, String name, String description, long defaultFoodMeasurementUnitId,
                            float defaultServingSize, FoodFormType formType,
                            NutritionalValuesEntry nutritionalValuesEntry) throws FitbitAPIException {
         setAccessToken(localUser);
